Changed the state structure in the movement states to better use the hierarchy of the State Machine
The Idle and Move state now work for both the airborne and grounded states.
Jump state has been removed as a concept for now, will see if it's needed later.

Added a struct for Movement values to make it easier to group values to together and simplify changing them with state changes.